+++ packages/frontend/src/ee/features/embed/EmbedExplore/components/EmbedExplore.tsx
@@ -21,7 +21,7 @@ const themeOverride: MantineThemeOverride = {
 
 const getInitialState = (exploreId: string, savedChart: SavedChart) => ({
     parameters: {},
-    shouldFetchResults: true,
+    parameterReferences: [],
     expandedSections: [
         ExplorerSection.FILTERS,
         ExplorerSection.VISUALIZATION,