+++ packages/frontend/src/pages/MinimalSavedExplorer.tsx
@@ -105,7 +105,9 @@ const MinimalSavedExplorer: FC = () => {
             initialState={
                 data
                     ? {
-                          shouldFetchResults: true,
+                          parameterReferences: Object.keys(
+                              data.parameters ?? {},
+                          ),
                           expandedSections: [ExplorerSection.VISUALIZATION],
                           unsavedChartVersion: {
                               tableName: data.tableName,